+++ packages/clerk-js/src/ui/contexts/components/Plans.tsx
@@ -99,7 +99,6 @@ export const usePlans = () => {
 type HandleSelectPlanProps = {
   plan: CommercePlanResource;
   planPeriod: CommerceSubscriptionPlanPeriod;
-  onSubscriptionChange?: () => void;
   mode?: 'modal' | 'mounted';
   event?: React.MouseEvent<HTMLElement>;
   appearance?: Appearance;
@@ -281,6 +280,10 @@ export const usePlansContext = () => {
   );
 
   const captionForSubscription = useCallback((subscription: CommerceSubscriptionResource) => {
+    if (subscription.pastDueAt) {
+      return localizationKeys('badge__pastDueAt', { date: subscription.pastDueAt });
+    }
+
     if (subscription.status === 'upcoming') {
       return localizationKeys('badge__startsAt', { date: subscription.periodStartDate });
     }
@@ -294,55 +297,44 @@ export const usePlansContext = () => {
     return;
   }, []);
 
+  const openSubscriptionDetails = useCallback(
+    (event?: React.MouseEvent<HTMLElement>) => {
+      const portalRoot = getClosestProfileScrollBox('modal', event);
+      clerk.__internal_openSubscriptionDetails({
+        for: subscriberType,
+        onSubscriptionCancel: () => {
+          revalidateAll();
+        },
+        portalRoot,
+      });
+    },
+    [clerk, subscriberType, revalidateAll],
+  );
+
   // handle the selection of a plan, either by opening the subscription details or checkout
   const handleSelectPlan = useCallback(
-    ({
-      plan,
-      planPeriod,
-      onSubscriptionChange,
-      mode = 'mounted',
-      event,
-      appearance,
-      newSubscriptionRedirectUrl,
-    }: HandleSelectPlanProps) => {
-      const subscription = activeOrUpcomingSubscriptionWithPlanPeriod(plan, planPeriod);
-
+    ({ plan, planPeriod, mode = 'mounted', event, appearance, newSubscriptionRedirectUrl }: HandleSelectPlanProps) => {
       const portalRoot = getClosestProfileScrollBox(mode, event);
 
-      if (subscription && subscription.planPeriod === planPeriod && !subscription.canceledAtDate) {
-        clerk.__internal_openPlanDetails({
-          plan,
-          initialPlanPeriod: planPeriod,
-          subscriberType,
-          onSubscriptionCancel: () => {
-            revalidateAll();
-            onSubscriptionChange?.();
-          },
-          appearance,
-          portalRoot,
-        });
-      } else {
-        clerk.__internal_openCheckout({
-          planId: plan.id,
-          // if the plan doesn't support annual, use monthly
-          planPeriod: planPeriod === 'annual' && plan.annualMonthlyAmount === 0 ? 'month' : planPeriod,
-          subscriberType,
-          onSubscriptionComplete: () => {
-            revalidateAll();
-            onSubscriptionChange?.();
-          },
-          onClose: () => {
-            if (session?.id) {
-              void clerk.setActive({ session: session.id });
-            }
-          },
-          appearance,
-          portalRoot,
-          newSubscriptionRedirectUrl,
-        });
-      }
+      clerk.__internal_openCheckout({
+        planId: plan.id,
+        // if the plan doesn't support annual, use monthly
+        planPeriod: planPeriod === 'annual' && plan.annualMonthlyAmount === 0 ? 'month' : planPeriod,
+        subscriberType,
+        onSubscriptionComplete: () => {
+          revalidateAll();
+        },
+        onClose: () => {
+          if (session?.id) {
+            void clerk.setActive({ session: session.id });
+          }
+        },
+        appearance,
+        portalRoot,
+        newSubscriptionRedirectUrl,
+      });
     },
-    [clerk, revalidateAll, activeOrUpcomingSubscription, subscriberType, session?.id],
+    [clerk, revalidateAll, subscriberType, session?.id],
   );
 
   const defaultFreePlan = useMemo(() => {
@@ -355,6 +347,7 @@ export const usePlansContext = () => {
     activeOrUpcomingSubscriptionBasedOnPlanPeriod: activeOrUpcomingSubscriptionWithPlanPeriod,
     isDefaultPlanImplicitlyActiveOrUpcoming,
     handleSelectPlan,
+    openSubscriptionDetails,
     buttonPropsForPlan,
     canManageSubscription,
     captionForSubscription,