+++ packages/clerk-js/src/ui/elements/Card/CardFooter.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import { colors } from '@/ui/utils/colors';
-
 import { useEnvironment } from '../../contexts';
 import { descriptors, Flex, Link, localizationKeys, useAppearance } from '../../customizables';
 import { useDevMode } from '../../hooks/useDevMode';
@@ -33,7 +31,7 @@ export const CardFooter = React.forwardRef<HTMLDivElement, CardFooterProps>((pro
       padding: `${t.space.$4} ${t.space.$8}`,
       borderTopWidth: t.borderWidths.$normal,
       borderTopStyle: t.borderStyles.$solid,
-      borderTopColor: t.colors.$neutralAlpha100,
+      borderTopColor: t.colors.$borderAlpha100,
     },
   });
 
@@ -49,12 +47,9 @@ export const CardFooter = React.forwardRef<HTMLDivElement, CardFooterProps>((pro
       elementDescriptor={descriptors.footer}
       sx={[
         t => ({
-          marginTop: `-${t.space.$2}`,
+          marginTop: `calc(${t.space.$2} * -1)`,
           paddingTop: t.space.$2,
-          background: common.mergedColorsBackground(
-            colors.setAlpha(t.colors.$colorBackground, 1),
-            t.colors.$neutralAlpha50,
-          ),
+          background: common.mutedBackground(t),
           '&:empty': {
             padding: 0,
             marginTop: 0,